requirements.txt
opencv-python
pillow
pytesseract
bunch
behave
pyhamcrest
